add BER, BPSK/QPSK decoder

+++ lib/SER_impl.cc
@@ -68,8 +68,8 @@ namespace gr {
     void
     SER_impl::forecast (int noutput_items, gr_vector_int &ninput_items_required)
     {
-      unsigned ninputs = ninput_items_required.size ();
-      for(unsigned i = 0; i < ninputs; i++)
+      unsigned long ninputs = ninput_items_required.size ();
+      for(unsigned long i = 0; i < ninputs; i++)
         ninput_items_required[i] = noutput_items;
     }
